requirements.txt
pytest~=7.4.3
selene~=2.0.0rc9
allure-pytest
selenium==4.29.0
allure-python-commons
python-dotenv==1.0.1
